hybrid-agent/requirements.txt
torch>=2.0.0
transformers>=4.30.0
fastapi>=0.95.0
uvicorn>=0.22.0
supabase>=1.0.3
pgvector>=0.2.0
PyMuPDF>=1.22.0
python-docx>=0.8.11
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0
jose>=1.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
nltk>=3.8.1
tenacity>=8.2.2